Various fixes

+++ lib/moku/filesystem.rb
@@ -82,6 +82,8 @@ module Moku
     # @param [Pathname] path
     def rm_empty_tree(path)
       raise ArgumentError, "path cannot be a non-directory file" if path.file?
+      return unless path.exist?
+      return unless path.empty?
 
       FileUtils.rmdir path, parents: true
     end
